+++ resharper/resharper-unity/src/Rider/UnityController.cs
@@ -1,9 +1,8 @@
 using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using Debugger.Common.ManagedSymbols;
 using JetBrains.Annotations;
-using JetBrains.Application.Threading;
+using JetBrains.Application.Components;
 using JetBrains.Application.Threading.Tasks;
 using JetBrains.Collections.Viewable;
 using JetBrains.Core;
@@ -14,43 +13,38 @@ using JetBrains.ReSharper.Host.Features;
 using JetBrains.ReSharper.Host.Features.Unity;
 using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
 using JetBrains.ReSharper.Plugins.Unity.Rider.UnitTesting;
-using JetBrains.ReSharper.Psi.Impl.CodeStyle;
 using JetBrains.ReSharper.UnitTestFramework.Strategy;
 using JetBrains.Rider.Model;
-using JetBrains.Threading;
 using JetBrains.Util;
-using JetBrains.Util.Special;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Rider
 {
     [SolutionComponent]
-    public class UnityController : IUnityController
+    public class UnityController : IUnityController, IHideImplementation<DefaultUnityController>
     {
-        private static readonly TimeSpan outUnityConnectionTimeout = TimeSpan.FromMinutes(1);
-        private static readonly string outUnityTimeoutMessage = $"Unity hasn't connected. Timeout {outUnityConnectionTimeout.TotalMilliseconds} ms is over.";
+        private static readonly TimeSpan ourUnityConnectionTimeout = TimeSpan.FromMinutes(10);
+        private static readonly string ourUnityTimeoutMessage = $"Unity hasn't connected. Timeout {ourUnityConnectionTimeout.TotalMilliseconds} ms is over.";
         private readonly UnityEditorProtocol myUnityEditorProtocol;
+        private readonly UnityVersion myUnityVersion;
         private readonly ISolution mySolution;
         private readonly Lifetime myLifetime;
-        private readonly UnitySolutionTracker mySolutionTracker;
-        private readonly IThreading myThreading;
         private readonly RdUnityModel myRdUnityModel;
 
         private FileSystemPath EditorInstanceJsonPath => mySolution.SolutionDirectory.Combine("Library/EditorInstance.json");
 
-        public UnityController(UnityEditorProtocol unityEditorProtocol, 
+        public UnityController(Lifetime lifetime, 
                                ISolution solution,
-                               Lifetime lifetime, 
-                               UnitySolutionTracker solutionTracker,
-                               IThreading threading)
+                               UnityEditorProtocol unityEditorProtocol,
+                               UnityVersion unityVersion)
+
         {
             if (solution.GetData(ProjectModelExtensions.ProtocolSolutionKey) == null)
                 return;
-            
+
             myUnityEditorProtocol = unityEditorProtocol;
+            myUnityVersion = unityVersion;
             mySolution = solution;
             myLifetime = lifetime;
-            mySolutionTracker = solutionTracker;
-            myThreading = threading;
             myRdUnityModel = solution.GetProtocolSolution().GetRdUnityModel();
         }
 
@@ -67,7 +61,11 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
                 return Task.FromResult(new ExitUnityResult(false, "No connection to Unity Editor.", null));
             }
 
-            var protocolTask = myUnityEditorProtocol.UnityModel.Value.ExitUnity.Start(lifetimeDef.Lifetime, Unit.Instance).AsTask();
+            var protocolTaskSource = new TaskCompletionSource<bool>();
+            mySolution.Locks.Tasks.StartNew(lifetimeDef.Lifetime, Scheduling.MainGuard, () => myUnityEditorProtocol.UnityModel.Value.ExitUnity.Start(lifetimeDef.Lifetime, Unit.Instance)
+                .AsTask());
+            var protocolTask = protocolTaskSource.Task;
+
             var waitTask = Task.WhenAny(protocolTask, Task.Delay(TimeSpan.FromSeconds(0.5), lifetimeDef.Lifetime)); // continue on timeout
             return waitTask.ContinueWith(t =>
             {
@@ -82,13 +80,13 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
         private Task<ExitUnityResult> WaitModelUpdate()
         {
             var successExitResult = new ExitUnityResult(true, null, null);
-            if (!myUnityEditorProtocol.UnityModel.HasValue()) 
+            if (!myUnityEditorProtocol.UnityModel.HasValue())
                 return Task.FromResult(successExitResult);
-            
+
             var taskSource = new TaskCompletionSource<ExitUnityResult>();
             var waitLifetimeDef = myLifetime.CreateNested();
             waitLifetimeDef.SynchronizeWith(taskSource);
-                        
+
             // Wait RdModel Update
             myUnityEditorProtocol.UnityModel.ViewNull(waitLifetimeDef.Lifetime, _ => taskSource.SetResult(successExitResult));
             return taskSource.Task;
@@ -101,7 +99,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
             {
                 if (model.UnityProcessId.HasValue())
                 {
-                    return model.UnityProcessId.Value;    
+                    return model.UnityProcessId.Value;
                 }
             }
             // no protocol connection - try to fallback to EditorInstance.json
@@ -119,12 +117,13 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
                 lifetimeDef.Lifetime,
                 (lt, model) => model.UnityProcessId.Advise(lt, id => source.TrySetResult(id)));
 
-            Task.Delay(outUnityConnectionTimeout, lifetimeDef.Lifetime).ContinueWith(_ =>
+            // ToDo Replace timeout with CancellationToken
+            Task.Delay(ourUnityConnectionTimeout, lifetimeDef.Lifetime).ContinueWith(_ =>
             {
                 if (source.Task.Status != TaskStatus.RanToCompletion)
-                    source.TrySetException(new TimeoutException(outUnityTimeoutMessage));
+                    source.TrySetException(new TimeoutException(ourUnityTimeoutMessage));
             }, lifetimeDef.Lifetime);
-            
+
             return source.Task;
         }
 
@@ -132,12 +131,15 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
         public string[] GetUnityCommandline()
         {
             var unityPathData = myRdUnityModel.UnityApplicationData;
-            if (!unityPathData.HasValue()) 
+            if (!unityPathData.HasValue())
                 return null;
             var unityPath = unityPathData.Value?.ApplicationPath;
             if (unityPath != null && PlatformUtil.RuntimePlatform == PlatformUtil.Platform.MacOsX)
                 unityPath = FileSystemPath.Parse(unityPath).Combine("Contents/MacOS/Unity").FullPath;
-            return unityPath != null ? new[] {unityPath, "-projectPath", mySolution.SolutionDirectory.FullPath} : null;
+
+            return unityPath == null
+                ? null 
+                : new[] { CommandLineUtil.QuoteIfNeeded(unityPath), "-projectPath", CommandLineUtil.QuoteIfNeeded(mySolution.SolutionDirectory.FullPath) };
         }
 
         public bool IsUnityGeneratedProject(IProject project)
@@ -145,9 +147,11 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
             return project.IsUnityGeneratedProject();
         }
 
-        public bool IsUnityEditorUnitTestRunStrategy(IUnitTestRunStrategy strategy)
+        public bool IsUnityEditorUnitTestRunStrategy(IUnitTestRunStrategy strategy) => strategy is RunViaUnityEditorStrategy;
+
+        public Version GetUnityVersion()
         {
-            return UnityNUnitServiceProvider.IsUnityUnitTestStrategy(mySolutionTracker, myRdUnityModel, myUnityEditorProtocol);
+            return myUnityVersion.ActualVersionForSolution.Value;
         }
 
         private ExitUnityResult KillProcess()