+++ resharper/resharper-unity/test/data/CSharp/Intentions/QuickFixes/RedundantEventFunction/Availability/Test02.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+public class MyBehaviour : MonoBehaviour
+{
+    public void DoSomething()
+    {
+        Update();
+    }
+
+    public void U{caret}pdate()
+    {
+    }
+}