+++ resharper/resharper-unity/test/src/CSharp/Intentions/QuickFixes/RedundantEventFunctionQuickFixTests.cs
@@ -1,0 +1,25 @@
+using JetBrains.ReSharper.FeaturesTestFramework.Intentions;
+using JetBrains.ReSharper.Intentions.QuickFixes.UsageChecking;
+using NUnit.Framework;
+
+namespace JetBrains.ReSharper.Plugins.Unity.Tests.CSharp.Intentions.QuickFixes
+{
+    [TestUnity]
+    public class RedundantEventFunctionQuickFixAvailabilityTests : QuickFixAvailabilityTestBase
+    {
+        protected override string RelativeTestDataPath => @"CSharp\Intentions\QuickFixes\RedundantEventFunction\Availability";
+
+        [Test] public void Test01() { DoNamedTest(); }
+        [Test] public void Test02() { DoNamedTest(); }
+        [Test] public void Test03() { DoNamedTest(); }
+    }
+
+    [TestUnity]
+    public class RedundantEventFunctionQuickFixTests : CSharpQuickFixTestBase<RemoveUnusedElementFix>
+    {
+        protected override string RelativeTestDataPath => @"CSharp\Intentions\QuickFixes\RedundantEventFunction";
+
+        [Test] public void Test01() { DoNamedTest(); }
+        [Test] public void Test02() { DoNamedTest(); }
+    }
+}