+++ rider/protocol/src/main/kotlin/model/editorPlugin/EditorPluginModel.kt
@@ -125,6 +125,18 @@ object EditorPluginModel: Root() {
         field("applicationVersion", string)
     }
 
+    val RunMethodData = structdef{
+        field("assemblyName", string)
+        field("typeName", string)
+        field("methodName", string)
+    }
+
+    val RunMethodResult =  classdef{
+        field("success", bool)
+        field("message", string)
+        field("stackTrace", string)
+    }
+
     init {
         property("play", bool)
         property("pause", bool)
@@ -153,11 +165,14 @@ object EditorPluginModel: Root() {
         property("unitTestLaunch", UnitTestLaunch)
         call("runUnitTestLaunch", void, bool)
 
+        call("runMethodInUnity", RunMethodData, RunMethodResult)
+
         property("editorLogPath", string)
         property("playerLogPath", string)
 
         property("ScriptCompilationDuringPlay", int)
-        sink("clearOnPlay", long)
+        property("lastPlayTime", long)
+        property("lastInitTime", long)
 
         call("generateUIElementsSchema", void, bool)
         call("exitUnity", void, bool)