+++ rider/protocol/src/main/kotlin/model/rider/RdUnityModel.kt
@@ -40,6 +40,18 @@ object RdUnityModel : Ext(SolutionModel.Solution) {
         field("requiresRiderPackage", bool)
     }
 
+    val RunMethodData = structdef{
+        field("assemblyName", string)
+        field("typeName", string)
+        field("methodName", string)
+    }
+
+    val RunMethodResult =  classdef{
+        field("success", bool)
+        field("message", string)
+        field("stackTrace", string)
+    }
+
     private val shaderInternScope = internScope()
 
     private val shaderContextDataBase = baseclass {
@@ -71,16 +83,19 @@ object RdUnityModel : Ext(SolutionModel.Solution) {
 
         property("unityApplicationData", UnityApplicationData)
 
+        call("runMethodInUnity", RunMethodData, RunMethodResult)
+
         property("editorLogPath", string)
         property("playerLogPath", string)
 
         property("play", bool)
-        sink("clearOnPlay", long)
         property("pause", bool)
-
         source("step", void)
         source("refresh", bool)
         source("showPreferences", void)
+
+        property("lastPlayTime", long)
+        property("lastInitTime", long)
 
         property("sessionInitialized", bool)
 
@@ -127,6 +142,8 @@ object RdUnityModel : Ext(SolutionModel.Solution) {
         field("backendSettings", aggregatedef("BackendSettings") {
             property("enableDebuggerExtensions", bool)
         })
+
+        property("riderFrontendTests", bool)
 
 
         call("requestShaderContexts", EditableEntityId, immutableList(shaderContextDataBase))