+++ rider/testData/integrationTests/DebuggerTest/checkExceptionBreakpointWithJustMyCode/source/NewBehaviourScript.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+public class NewBehaviourScript : MonoBehaviour
+{
+    void Update()
+    {
+         throw new System.ArgumentException();
+    }
+}